Render markup in selector options text
This change allows Selectize to correctly render any markup found in the `text` value of the options.

The `text` property of the Selectize options is passed through `htmlDecode()`. This function attempts to decode any encoded HTML, otherwise returning original string unchanged.

All Selectize `options` in the are also updated with a `search` property. This is assigned just the `.text()` value from the `text` property (after unescaping if needed), so that typing `div` or `img` in the search field won't match any markup that may have been present.